requirements.txt
feedparser==6.0.11
Jinja2==3.1.6
PyYAML==6.0.2
requests==2.32.4
pytest==8.4.2
